+++ assets/js/main.js
@@ -2,7 +2,7 @@
 // import "../../node_modules/popper.js/dist/popper.min.js"
 import "../../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js";
 
-import { FoodList, selectList } from './menu/FoodsList.js'
+import { FoodList } from './menu/FoodsList.js'
 import { NavBar } from "./NavBar.js";
 import * as UserManager from './auth/UserManager.js'; 
 import { CheckForUser } from "./auth/CheckForUser.js";
@@ -22,12 +22,24 @@ const headerElement = document.querySelector("header");
         if (event.target.id === "logout") {
           UserManager.logoutUser();
           NavBar();
-          selectList();
-          FoodList();
         }
     })
 
 
+
+    contentElement.addEventListener("click", event => {
+      if (event.target.id === "login__submit"){
+        const userObject = {
+          name: document.querySelector("#login_name").value,
+          email: document.querySelector("#login_email").value,
+        }
+        UserManager.loginUser(userObject).then(() => {
+          startSO();
+        })
+       
+      }
+  
+    })
 
 ///////////// end event listeners /////////////////////
